+++ src/Contracts/Filter.php
@@ -4,5 +4,5 @@ namespace Vdhicts\Cyberfusion\ClusterApi\Contracts;
 
 interface Filter
 {
-    public function toArray(): array;
+    public function toQuery(): string;
 }